Improved Examples Part 1:refactor material modals
